Conflicts:
	nts.uk/uk.com/build.gradle

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.app/src/main/java/nts/uk/ctx/at/record/app/find/standardtime/dto/AgreementYearSettingDto.java
@@ -1,7 +1,5 @@
 package nts.uk.ctx.at.record.app.find.standardtime.dto;
 
-import java.math.BigDecimal;
-
 import lombok.Data;
 
 @Data
@@ -9,7 +7,7 @@ public class AgreementYearSettingDto {
 
 	private int yearValue;
 	
-	private Long errorOneYear;
+	private int errorOneYear;
 	
-	private Long alarmOneYear;
+	private int alarmOneYear;
 }
